CovidManager/requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.7
django-crispy-forms==1.12.0
django-grappelli==2.15.1
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==19.9.0
psycopg2==2.9.1
pytz==2021.1
sqlparse==0.4.2
typing-extensions==3.10.0.2
whitenoise==5.3.0
